Metrics and logging.
The Go attitude seems to be 'logging responses is an application problem, not a server problem'.
After a brief sulk I agree with them.  This is the simplest approach I could come up with.
All others involve hijacking the http response or reflection.  This is the sort of magic I don't
want in my life anymore.

Resolves #11.